+++ crates/teloxide-core/src/adaptors/erased.rs
@@ -214,9 +214,11 @@ where
         edit_message_live_location_inline,
         stop_message_live_location,
         stop_message_live_location_inline,
+        edit_message_checklist,
         send_venue,
         send_contact,
         send_poll,
+        send_checklist,
         send_dice,
         send_chat_action,
         set_message_reaction,
@@ -496,6 +498,14 @@ trait ErasableRequester<'a> {
         inline_message_id: String,
     ) -> ErasedRequest<'a, StopMessageLiveLocationInline, Self::Err>;
 
+    fn edit_message_checklist(
+        &self,
+        business_connection_id: BusinessConnectionId,
+        chat_id: ChatId,
+        message_id: MessageId,
+        checklist: InputChecklist,
+    ) -> ErasedRequest<'a, EditMessageChecklist, Self::Err>;
+
     fn send_venue(
         &self,
         chat_id: Recipient,
@@ -518,6 +528,13 @@ trait ErasableRequester<'a> {
         question: String,
         options: Vec<InputPollOption>,
     ) -> ErasedRequest<'a, SendPoll, Self::Err>;
+
+    fn send_checklist(
+        &self,
+        business_connection_id: BusinessConnectionId,
+        chat_id: ChatId,
+        checklist: InputChecklist,
+    ) -> ErasedRequest<'a, SendChecklist, Self::Err>;
 
     fn send_dice(&self, chat_id: Recipient) -> ErasedRequest<'a, SendDice, Self::Err>;
 
@@ -1428,6 +1445,23 @@ where
         Requester::stop_message_live_location_inline(self, inline_message_id).erase()
     }
 
+    fn edit_message_checklist(
+        &self,
+        business_connection_id: BusinessConnectionId,
+        chat_id: ChatId,
+        message_id: MessageId,
+        checklist: InputChecklist,
+    ) -> ErasedRequest<'a, EditMessageChecklist, Self::Err> {
+        Requester::edit_message_checklist(
+            self,
+            business_connection_id,
+            chat_id,
+            message_id,
+            checklist,
+        )
+        .erase()
+    }
+
     fn send_venue(
         &self,
         chat_id: Recipient,
@@ -1455,6 +1489,15 @@ where
         options: Vec<InputPollOption>,
     ) -> ErasedRequest<'a, SendPoll, Self::Err> {
         Requester::send_poll(self, chat_id, question, options).erase()
+    }
+
+    fn send_checklist(
+        &self,
+        business_connection_id: BusinessConnectionId,
+        chat_id: ChatId,
+        checklist: InputChecklist,
+    ) -> ErasedRequest<'a, SendChecklist, Self::Err> {
+        Requester::send_checklist(self, business_connection_id, chat_id, checklist).erase()
     }
 
     fn send_dice(&self, chat_id: Recipient) -> ErasedRequest<'a, SendDice, Self::Err> {